refactor animatedText
